Update imports to accommodate relocation of 'bcl2fastq_utils.py' to 'bcl2fastq/utils.py'.
