+++ auto_process_ngs/barcodes/pipeline.py
@@ -29,9 +29,9 @@ import shutil
 import tempfile
 from ..analysis import AnalysisFastq
 from ..applications import Command
-from ..bcl2fastq_utils import get_nmismatches
-from ..bcl2fastq_utils import bases_mask_is_valid
-from ..bcl2fastq_utils import check_barcode_collisions
+from ..bcl2fastq.utils import get_nmismatches
+from ..bcl2fastq.utils import bases_mask_is_valid
+from ..bcl2fastq.utils import check_barcode_collisions
 from ..pipeliner import Pipeline
 from ..pipeliner import PipelineTask
 from ..pipeliner import PipelineCommandWrapper
