+++ auto_process_ngs/bcl2fastq/pipeline.py
@@ -63,14 +63,14 @@ from ..applications import general as general_apps
 from ..applications import bcl2fastq as bcl2fastq_apps
 from ..applications import Command
 from ..barcodes.pipeline import AnalyseBarcodes
-from ..bcl2fastq_utils import available_bcl2fastq_versions
-from ..bcl2fastq_utils import bases_mask_is_valid
-from ..bcl2fastq_utils import bcl_to_fastq_info
-from ..bcl2fastq_utils import check_barcode_collisions
-from ..bcl2fastq_utils import get_bases_mask
-from ..bcl2fastq_utils import get_nmismatches
-from ..bcl2fastq_utils import get_sequencer_platform
-from ..bcl2fastq_utils import make_custom_sample_sheet
+from ..bcl2fastq.utils import available_bcl2fastq_versions
+from ..bcl2fastq.utils import bases_mask_is_valid
+from ..bcl2fastq.utils import bcl_to_fastq_info
+from ..bcl2fastq.utils import check_barcode_collisions
+from ..bcl2fastq.utils import get_bases_mask
+from ..bcl2fastq.utils import get_nmismatches
+from ..bcl2fastq.utils import get_sequencer_platform
+from ..bcl2fastq.utils import make_custom_sample_sheet
 from ..fastq_utils import group_fastqs_by_name
 from ..fileops import Location
 from ..icell8.utils import get_bases_mask_icell8
