+++ auto_process_ngs/commands/setup_cmd.py
@@ -16,8 +16,8 @@ except ImportError:
     # Failed to get Python3 urlopen, fallback to Python2
     from urllib2 import urlopen
     from urllib2 import URLError
-from ..bcl2fastq_utils import get_sequencer_platform
-from ..bcl2fastq_utils import make_custom_sample_sheet
+from ..bcl2fastq.utils import get_sequencer_platform
+from ..bcl2fastq.utils import make_custom_sample_sheet
 from ..applications import general as general_applications
 from ..fileops import exists
 from ..fileops import Location
